Move registry into a separate package
This makes it possible for the registry-related tests to run in a separate
process, so that registering drivers for executable tests won't fail in
strange ways.

A side effect is that now `TestNew`'s test registrations are permanent (as
permanent as the test process, anyway), but that shouldn't matter much.
