+++ internal/registry/registry.go
@@ -1,4 +1,6 @@
-package kivik
+// Package registry handles driver registrations. It's in a separate package
+// to facilitate testing.
+package registry
 
 import (
 	"sync"
@@ -23,4 +25,12 @@ func Register(name string, driver driver.Driver) {
 		panic("kivik: Register called twice for driver " + name)
 	}
 	drivers[name] = driver
+}
+
+// Driver returns the driver registered with the requested name, or nil if
+// it has not been registered.
+func Driver(name string) driver.Driver {
+	driversMu.RLock()
+	defer driversMu.RUnlock()
+	return drivers[name]
 }
